requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
distro==1.9.0
fastapi==0.115.9
fire==0.7.0
fonttools==4.56.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.48.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imgaug==0.4.0
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
matplotlib==3.10.1
networkx==3.4.2
numpy==1.26.4
openai==1.65.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle-gpu==2.6.0.post120
pdf2image==1.17.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==3.20.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
PyWavelets==1.8.0
PyYAML==6.0.2
RapidFuzz==3.12.1
requests==2.32.3
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
